+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Prevent errors when `status` field in cockroach_cluster resource drifts due to
+  external changes.
+
 ## [1.14.0] - 2025-09-03
 
 ### Added